requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=2.1.0=pypi_0
aiohappyeyeballs=2.6.1=pypi_0
aiohttp=3.11.14=pypi_0
aiosignal=1.3.2=pypi_0
attrs=25.3.0=pypi_0
bzip2=1.0.8=h5eee18b_6
ca-certificates=2025.2.25=h06a4308_0
certifi=2025.1.31=pypi_0
charset-normalizer=3.4.1=pypi_0
contourpy=1.3.1=pypi_0
cycler=0.12.1=pypi_0
expat=2.6.4=h6a678d5_0
filelock=3.18.0=pypi_0
fonttools=4.56.0=pypi_0
frozenlist=1.5.0=pypi_0
fsspec=2025.3.0=pypi_0
grpcio=1.71.0=pypi_0
idna=3.10=pypi_0
iniconfig=2.0.0=pypi_0
jinja2=3.1.6=pypi_0
joblib=1.4.2=pypi_0
kiwisolver=1.4.8=pypi_0
lazy-loader=0.4=pypi_0
ld_impl_linux-64=2.40=h12ee557_0
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
lightning-utilities=0.14.1=pypi_0
markdown=3.7=pypi_0
markupsafe=3.0.2=pypi_0
matplotlib=3.10.1=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.1.0=pypi_0
ncurses=6.4=h6a678d5_0
networkx=3.4.2=pypi_0
numpy=2.2.4=pypi_0
openssl=3.0.16=h5eee18b_0
packaging=24.2=pypi_0
pandas=2.2.3=pypi_0
patsy=1.0.1=pypi_0
pillow=11.1.0=pypi_0
pip=25.0=py312h06a4308_0
pluggy=1.5.0=pypi_0
propcache=0.3.0=pypi_0
protobuf=6.30.1=pypi_0
pyparsing=3.2.1=pypi_0
pytest=8.3.5=pypi_0
python=3.12.9=h5148396_0
python-dateutil=2.9.0.post0=pypi_0
pytorch-lightning=2.5.0.post0=pypi_0
pytz=2025.1=pypi_0
pyyaml=6.0.2=pypi_0
readline=8.2=h5eee18b_0
regex=2024.11.6=pypi_0
requests=2.32.3=pypi_0
scikit-learn=1.6.1=pypi_0
scipy=1.15.2=pypi_0
setuptools=75.8.0=py312h06a4308_0
six=1.17.0=pypi_0
sqlite=3.45.3=h5eee18b_0
statsmodels=0.14.4=pypi_0
tensorboard=2.19.0=pypi_0
tensorboard-data-server=0.7.2=pypi_0
threadpoolctl=3.6.0=pypi_0
tk=8.6.14=h39e8969_0
torchmetrics=1.6.3=pypi_0
tqdm=4.67.1=pypi_0
typing-extensions=4.12.2=pypi_0
tzdata=2025.1=pypi_0
urllib3=2.3.0=pypi_0
werkzeug=3.1.3=pypi_0
wheel=0.45.1=py312h06a4308_0
xz=5.6.4=h5eee18b_1
yarl=1.18.3=pypi_0
zlib=1.2.13=h5eee18b_1
